Login iPad without link
If you use this module with iPad, the login page contains links that are not allowed by Apple, threatening to have rejected the app.
But we can add "display = mobile" in the query params of the /oauth/authorize url to force it to display the mobile oauth flow. In this way, it is as if we access with the iPhone.
I already tried, and is fully compatible with the module.